Fix x64 breakage.
BUG=

Review URL: https://codereview.chromium.org/144263002

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@18723 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/mark-compact.cc
@@ -3078,7 +3078,7 @@ void MarkCompactCollector::EvacuatePages() {
     // crash in concurrent sweeping.
     CHECK(p->IsEvacuationCandidate() ||
           p->IsFlagSet(Page::RESCAN_ON_EVACUATION));
-    CHECK_EQ(p->parallel_sweeping(), 0);
+    CHECK_EQ(static_cast<int>(p->parallel_sweeping()), 0);
     if (p->IsEvacuationCandidate()) {
       // During compaction we might have to request a new page.
       // Check that space still have room for that.
